backend/requirements.txt
# Core dependencies
discord.py==2.3.1
python-dotenv==1.0.0
aiomysql==0.2.0
aiohttp==3.9.4

# Logging
colorlog==6.7.0

# Optional dependencies
# requests==2.32.2  # If needed for other features

